requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
blinker==1.8.2
click==8.1.7
Flask==2.3.3
frozenlist==1.4.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
pyee==11.1.0
python-dotenv==1.0.1
slackclient==2.9.4
slackeventsapi==3.0.1
typing_extensions==4.11.0
Werkzeug==3.0.3
yarl==1.9.4
openai==1.30.1
langchain=0.2.0
langchain-core=0.2.0
langchain-community==0.2.0
langchain-text-splitters=0.2.0
langchain-openai=0.1.7
chromadb=0.5.0